Made the following changes: 1. Inside who_am_i.h - Changed the node to motor node from geo node 2. Inside can_handler.c - Uncommented can_driver__motor_steering_mia() in can_handler__handle_all_mia() function - Uncommented can_driver__decode_motor_steering() in can_handler__handle_all_incoming_messages() function 3. Inside test_can_handler.c -Added expects for can_driver__motor_steering_mia() and can_driver__decode_motor_steering() 4. Inside motor_wrapper.c - Added includes for can_driver_node.h and project.h - Tied the decoded_steering_value_from_driver with can_driver__get_driver_steering() function 5. Inside servo.* - In servo__steer_processor function, changed function parameter from int to int16_t

+++ projects/canster/l5_application/hardware/motor/servo/servo.c
@@ -43,7 +43,7 @@ void servo__steer_init(void) {
   servo__configure_pwm_channel();
 }
 
-void servo__steer_processor(int steering_value) {
+void servo__steer_processor(int16_t steering_value) {
   switch (steering_value) {
   case -2:
     servo__steer_hard_left();
