+++ projects/canster/l5_application/hardware/motor/servo/servo.h
@@ -1,4 +1,5 @@
 #pragma once
 
+#include <stdint.h>
 void servo__steer_init(void);
-void servo__steer_processor(int steering_value); // [-2,2]
+void servo__steer_processor(int16_t steering_value); // [-2,2]
